// src/types/generated/DataPack.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DataPackQuality } from "./DataPackQuality";

export type DataPack = { version: string, patch: string | null, region: string | null, sources: Array<string>, quality: DataPackQuality, fallback: boolean, confidence: string | null, generated_at: number | null, };